Sync mode fixes and status updates
This commit fixes a bug where the DRCluster objects are not created
properly for MetroDR due to absence of secrets for external mode.

As a temporary workaround, the region for the DRCluster will be empty in case of sync mode and needs to be updated by user.
This workaround needs to be there till external mode clusters are
supported throughout MCO.

The MirrorPeer phase has two new phases for sync mode, S3ProfileSynced
and S3ProfileSyncing.

Signed-off-by: Vineet Badrinath <vineetbnath@gmail.com>