Fix issue with empty s3 profile config
This commit fixes issue where s3 profile config is not updated in the Ramen config map.
This is due to a race condition existing between MCO and Ramen where Ramen starts after MCO during upgrade at which point s3profile is updated and  Ramen replaces this with newly created default config.

This fix will ensure that any changes to Ramen configmap is watched and triggers reconcile for all secrets in the current namespace which should ensure that s3profile is properly updated regardless of the upgrade sequence.

Signed-off-by: vbadrina <vbadrina@redhat.com>