Fetches cluster fsid through secret and not mc
This commit fixes the issue with hub recovery where the new hub cluster does not have managedcluster CRs updated due to clusterclaims
on the managed clusters being in an non updated state.

We are changing the method from looking into the MC CRs to fetching it from the rook secret synced on hub.

Signed-off-by: vbadrina <vbadrina@redhat.com>