migrate addons to controller-runtime
This commit merges token exchange addon and maintenance addon into a
single operator and migrates it to controller-runtime. This operator is
now considered a single addon and is deployed as such.

Due to this change reconciles are now driven via events and watches as
opposed to the previous periodic resync approach. These events are handled
much more efficiently and will provide better resource utilization and
faster event handling.

Signed-off-by: Umanga Chapagain <chapagainumanga@gmail.com>