remove automated configuration of MultiClusterService
MirrorPeer.Spec.OverlappingCIDR was added to allow users to automatically
enable StorageCluster.Spec.Netowrking.MultiClusterService when needed. But
this implementation had some limitations due to which it was not generally
used.

This commit now drops the changes completely to avoid unwanted usage. This
will be replaced by alternate solutions which are more reliable.

Signed-off-by: Umanga Chapagain <uchapaga@redhat.com>