remove deprecated functionalities
Since ODF has taken responsibility of configuring itself for Disaster
Recovery, we are removing all the functionalities that are now
implementation details of ODF. Details of the changes are as follows:

- Removed blue and green secret controller
We no longer exchange internal ODF secrets between cluster peers.

- Removed blue/green secret processing from hub
Since we no longer exchange secrets between managed clusters, we do
not require the machinery to reconcile these secrets on the hub.

- Removed maintenance mode controller
We no longer manage Ramen MaintenanceMode operation for ODF. This
is handled internally in ODF.

- Removed configuration of CSI Addons
We no longer manage CSI addons configuration.

- Removed setting StorageCluster.Spec.Mirroring
We no longer set this spec as it is deprecated. We use the new
StorageClusterPeer CR to define mirroring relationships.

- Removed ODF internal mode specific code as its not required anymore

- Removed references to mirrorpeer.spec.storageclusteref.namespace

Signed-off-by: Umanga Chapagain <uchapaga@redhat.com>