explicitly name the MirrorPeer controller on spoke
We have 2 MirrorPeer controllers - one runs on the hub and the other
runs on spoke cluster. If in case, same cluster is used as hub and spoke,
there are chances of one of these controllers failing to register itself
because it is considered a duplicate. So this commit explicitly sets a
name for the agent mirrorpeer controller to disambiguate.

Signed-off-by: Umanga Chapagain <uchapaga@redhat.com>