read configuration from envvar or dotenv file as needed
While running in-cluster deployments, some runtime parameters are
provided as environment variables via Kubernetes downward-API and
other schemes. But, while running out-of-cluster deployments, same
schemes may not be possible. For such cases, this commit adds a
capability to provide parameters as dotenv file. The dotenv file
path should be provided as an argument during deployment.

Running out-of-cluster is usually only preferred for developement
environments and should not be used for production.

Signed-off-by: Umanga Chapagain <uchapaga@redhat.com>