set currentNamespace on all controllers during startup
The current namespace in which the controller is running does not change.
So, it is enough to get the current namespace when the controllers are
starting and re-use it. This commit adds the current namespace as a
private field of the reconciler and initializes it during startup.
All subsequent functions that requires the current execution namespace
will receive it from the reconciler.

Signed-off-by: Umanga Chapagain <uchapaga@redhat.com>