add external mode odf config to odf-client-info ConfigMap
The odf-client-info is created for all ODF deployment types. But,
its data is only populated when ODF accepts client connections.
This commit changes that behavior to populate data even in case of
external mode deployments which does not accept client connections.
A hypothetical client is assumed and used to create a map key as one
does not exist for external mode.

Signed-off-by: Umanga Chapagain <uchapaga@redhat.com>