add a watcher for ramen-hub-operator-config cm
add a watcher for ramen-hib-operator-config cm, as it is owned
by ramen operator bundle/CSV and if ramen makes any change to it
then the cm is reset by the OLM during upgrades. So, to keep the cm
we should watch the cm and update it as and when needed.

Signed-off-by: Nikhil-Ladha <nikhilladha1999@gmail.com>