use odf-client-info cm to check peerRefType
Currently, getPeerRefType function checks for clientType
differently for a managedcluster vs hub cluster. In case,
of managed cluster it checks the odf-info cm present
in the managecluster and in such a case if we have clients
with different storageclusterref then it will return wrong
clientType.
Instead, use "odf-client-info" cm present on hub to check
for peerRefType of the managedclusters part of a
mirrorpeer such that if we have storageclients with different
storageclusterref we are able to verify the type correctly.

Signed-off-by: Nikhil-Ladha <nikhilladha1999@gmail.com>